agent_service/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
requests
pydantic
mcp
google-adk